refactor: replace SplashWaveEffect with WaveEffect and restructure related files
- Removed SplashWaveEffect component and its associated files.
- Introduced WaveEffect component with WebGL rendering for wave effects.
- Updated imports in AnimationSettingsPage and guideMain to use WaveEffect.
- Added new shader configurations and utility functions for WaveEffect.
