+++ src/renderer/guideMain.tsx
@@ -30,7 +30,7 @@ import { StrictMode } from 'react';
 import type { ReactElement } from 'react';
 import { createRoot } from 'react-dom/client';
 import './styles/guide.css';
-import { SplashWaveEffect } from './components/components/SplashWaveEffect';
+import { WaveEffect } from './components/components/WaveEffect';
 
 /** 引导窗口根组件（基础架构占位，后续扩展配置步骤） */
 function GuideApp(): ReactElement {
@@ -41,7 +41,7 @@ function GuideApp(): ReactElement {
 
   return (
     <div className="guide-container">
-      <SplashWaveEffect />
+      <WaveEffect />
       <div className="guide-content">
         <h1>eIsland 引导配置</h1>
         <button onClick={handleComplete}>完成配置</button>
